fix controller authorization error

+++ controllers/user-friend.js
@@ -1,23 +1,29 @@
 import UserFriendService from "../services/user-friend.js";
-import runController from "./runner.js";
+import {authorizeRequest, runController} from "./runner.js";
 
 // delete
 async function deleteUser(req, res) {
-    const deleteUserData = await UserFriendService.deleteUser(
-        req.params.username
-    );
+    let deleteUserData = authorizeRequest(req.user, req.params.username);
 
-    runController(req.user, req.params.username, deleteUserData, res);
+    if (!deleteUserData)
+        deleteUserData = await UserFriendService.deleteUser(
+            req.params.username
+        );
+
+    runController(deleteUserData, res);
 }
 
 // delete
 async function removeFriend(req, res) {
-    const removeFriendData = await UserFriendService.removeFriend(
-        req.params.username,
-        req.params.fusername
-    );
+    let removeFriendData = authorizeRequest(req.user, req.params.username);
 
-    runController(req.user, req.params.username, removeFriendData, res);
+    if (!removeFriendData)
+        removeFriendData = await UserFriendService.removeFriend(
+            req.params.username,
+            req.params.fusername
+        );
+
+    runController(removeFriendData, res);
 }
 
 // get
@@ -27,7 +33,7 @@ async function getUserFriends(req, res) {
         req.params.username
     );
 
-    runController("", "", getUserFriendsData, res);
+    runController(getUserFriendsData, res);
 }
 
 // post
@@ -37,17 +43,20 @@ async function sendFriendRequest(req, res) {
         req.params.username
     );
 
-    runController("", "", sendFriendRequestData, res);
+    runController(sendFriendRequestData, res);
 }
 
 // patch
 async function acceptRequest(req, res) {
-    const acceptRequestData = await UserFriendService.acceptRequest(
-        req.params.fusername,
-        req.params.username
-    );
+    let acceptRequestData = authorizeRequest(req.user, req.params.username);
 
-    runController(req.user, req.params.username, acceptRequestData, res);
+    if (!acceptRequestData)
+        acceptRequestData = await UserFriendService.acceptRequest(
+            req.params.fusername,
+            req.params.username
+        );
+
+    runController(acceptRequestData, res);
 }
 
 export default {deleteUser, removeFriend, getUserFriends, sendFriendRequest, acceptRequest};
